refactor: GameScreen 컴포넌트 분리
